add database freshness check
Add a flag to process.env to store the state of the database (whether
new blocks are being added or not). A message is sent to Slack whenever
the state changes. The check interval was increased to 70 seconds.
Slack channel and token are now stored in .env instead of config.

Add middleware that sends Internal Server Error response if new blocks
are not being added to the database.

Fix some flow errors that arose, mostly just add 'any' to function
arguments that had no type.

Add Cardano Explorer check to calculate the expected block number. This
way, the server can stop sending responses also in the case when the
block number did change but is too far behind the actual newest block.

Add comments explaining the way block number is checked and when the
database is considered up to date.