Setup pm2 to launch two instances
Setup pm2 to launch a main instance and a backup instance running on
different ports. If process.env.PORT is provided, the server will
listen on that port instead of the one specified in the server config.
Since the server config object is not writable, its port could not be
directly overwritten. Add an instance config file to avoid warnings.
The build script now deletes the old build and the start script doesn't
call build anymore. Add server port to custom environment variables.