Fix healthcheck (#21)
* Fix healthcheck

Move the delay to the end of the healthcheck loop so the first check
is done immediately instead of after 70 seconds. Add process name to
the Slack message.

* Fix healthcheck errors

Convert the block number fetched from the database to a number. Fix
a scope error when comparing the fetched number to the expected one.