Add bulk address summary endpoint
* Add bulk address summary endpoint

/api/bulk/addresses/summary accepts POST requests with arrays of
addresses. If the array is empty, too long or any of the addresses is
invalid, Left is returned. Otherwise, the addresses, the number of
transactions and the transaction data are returned. The transactions
are ordered by time, descending.

* Refactor db-api and legacy routes

Remove addressSummary function, replace the call to it with a call to
bulkAddressSummary with one element array as argument.